Refactor, move all route definitions to routes.js.
