--- src/authentication/handlers/index.js
@@ -1,8 +1,0 @@
-'use strict'
-
-module.exports = {
-  AuthCallbackRequest: require('./auth-callback-request'),
-  LoginConsentRequest: require('./login-consent-request'),
-  LogoutRequest: require('./logout-request'),
-  SelectProviderRequest: require('./select-provider-request')
-}
