--- src/authentication/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  oidc: require('./webid-oidc'),
-  forceUser: require('./force-user.js')
-}
